using port from client

+++ src/relay.h
@@ -25,7 +25,7 @@ class Relay
 	void log(std::string s);
 	std::string mMessage;
 
-	unsigned short mPort;
+	int mPort;
 
 
 };
